split common defs into types, kinds and operators
