+++ common/abstract-syntax-tree/ast_function.hpp
@@ -8,7 +8,7 @@
 #include "ast_parameter.hpp"
 #include "ast_stmt.hpp"
 #include "../token/token.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ast_visitor.hpp"
 
 namespace syntax::ast {
