+++ common/abstract-syntax-tree/ast_literal_expr.hpp
@@ -3,7 +3,7 @@
 
 #include "ast_expr.hpp"
 #include "../token/token.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ast_visitor.hpp"
 
 namespace syntax::ast {
