+++ common/abstract-syntax-tree/ast_variable_decl_stmt.hpp
@@ -6,7 +6,7 @@
 #include "ast_stmt.hpp"
 #include "ast_expr.hpp"
 #include "../token/token.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ast_visitor.hpp"
 
 namespace syntax::ast {
