+++ common/intermediate-representation-tree/ir_binary_expr.hpp
@@ -4,7 +4,7 @@
 #include <memory>
 
 #include "ir_expr.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ir_visitor.hpp"
 
 namespace ir {
