+++ common/intermediate-representation-tree/ir_function.hpp
@@ -8,7 +8,7 @@
 #include "ir_node.hpp"
 #include "ir_parameter.hpp"
 #include "ir_stmt.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ir_visitor.hpp"
 
 namespace ir {
