+++ common/intermediate-representation-tree/ir_literal_expr.hpp
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "ir_expr.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ir_visitor.hpp"
 
 namespace ir {
