+++ common/intermediate-representation-tree/ir_parameter.hpp
@@ -5,7 +5,7 @@
 #include <string_view>
 
 #include "ir_node.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ir_visitor.hpp"
 
 namespace ir {
