+++ common/intermediate-representation-tree/ir_temporary_expr.hpp
@@ -8,7 +8,7 @@
 
 #include "ir_node.hpp"
 #include "ir_expr.hpp"
-#include "../defs/defs.hpp"
+#include "../defs/types.hpp"
 #include "../visitor/ir_visitor.hpp"
 
 namespace ir {
